update cts v3
