+++ acceptance/openstack/cts/v3/trackers_test.go
@@ -1,19 +1,16 @@
 package v3
 
 import (
-	"os"
 	"testing"
 
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/clients"
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/openstack/cts"
+	"github.com/opentelekomcloud/gophertelekomcloud/openstack/common/pointerto"
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/cts/v3/tracker"
 	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
 )
 
 func TestTrackersLifecycle(t *testing.T) {
-	if os.Getenv("RUN_CTS_TRACKER") == "" {
-		t.Skip("unstable test")
-	}
 	client, err := clients.NewCTSV3Client()
 	th.AssertNoErr(t, err)
 
@@ -28,8 +25,10 @@ func TestTrackersLifecycle(t *testing.T) {
 		TrackerName:  "system",
 		IsLtsEnabled: true,
 		ObsInfo: tracker.ObsInfo{
-			BucketName:     bucketName,
-			FilePrefixName: "test-prefix",
+			BucketName:      bucketName,
+			FilePrefixName:  "test-prefix",
+			CompressType:    "json",
+			IsSortByService: pointerto.Bool(true),
 		},
 	})
 
@@ -46,6 +45,7 @@ func TestTrackersLifecycle(t *testing.T) {
 	th.AssertEquals(t, "enabled", ctsTracker.Status)
 	th.AssertEquals(t, false, *ctsTracker.ObsInfo.IsObsCreated)
 	th.AssertEquals(t, bucketName, ctsTracker.ObsInfo.BucketName)
+	th.AssertEquals(t, "json", ctsTracker.ObsInfo.CompressType)
 
 	t.Logf("Attempting to update CTSv3 Tracker: %s", ctsTracker.TrackerName)
 	ltsEnable := false
@@ -56,10 +56,11 @@ func TestTrackersLifecycle(t *testing.T) {
 		IsLtsEnabled: &ltsEnable,
 		ObsInfo: tracker.ObsInfo{
 			FilePrefixName: "test-2-",
+			CompressType:   "gzip",
 		},
 	})
 	th.AssertNoErr(t, err)
-	t.Logf("Updated CTSv1 Tracker: %s", ctsTracker.TrackerName)
+	t.Logf("Updated CTSv3 Tracker: %s", ctsTracker.TrackerName)
 
 	trackerList, err := tracker.List(client, ctsTracker.TrackerName)
 	th.AssertNoErr(t, err)
@@ -70,4 +71,6 @@ func TestTrackersLifecycle(t *testing.T) {
 	// if tracker is disabled LTS status can't be changed
 	th.AssertEquals(t, trackerGet.Lts.IsLtsEnabled, false)
 	th.AssertEquals(t, trackerGet.ObsInfo.FilePrefixName, "test-2-")
+	// update of `compress_type` doesn't work
+	// th.AssertEquals(t, "gzip", ctsTracker.ObsInfo.CompressType)
 }
